Add useTextExpose hook
