+++ app/components/Layout.tsx
@@ -1,7 +1,5 @@
 import clsx from 'clsx';
 import type {ReactNode} from 'react';
-import {useCallback, useEffect, useState, useRef} from 'react';
-import {useAnalytics} from '@shopify/hydrogen';
 import {PackTestProvider} from '@pack/hydrogen';
 
 import {Analytics} from '~/components/Analytics';
@@ -11,13 +9,13 @@ import {Header} from '~/components/Header';
 import {Modal} from '~/components/Modal';
 import {ProductModal} from '~/components/Product/ProductModal';
 import {Search} from '~/components/Search';
-import {AnalyticsEvent} from '~/components/Analytics/constants';
 import {usePreviewModeCustomerInit} from '~/lib/customer';
 import {
   useCartAddDiscountUrl,
   usePromobar,
   useScrollToHashOnNavigation,
   useSetViewportHeightCssVar,
+  useTestExpose,
 } from '~/hooks';
 export function Layout({children}: {children: ReactNode}) {
   const {mainPaddingTopClass} = usePromobar();
@@ -26,47 +24,7 @@ export function Layout({children}: {children: ReactNode}) {
   useScrollToHashOnNavigation();
   useSetViewportHeightCssVar();
 
-  const [hasUserConsent, setHasUserConsent] = useState<boolean>(false);
-  const {publish} = useAnalytics();
-
-  const publishRef = useRef(publish);
-
-  useEffect(() => {
-    publishRef.current = publish;
-  }, [publish]);
-
-  const handleTestExpose = (test: any) => {
-    if (hasUserConsent) {
-      setTimeout(() => {
-        publishRef?.current?.(AnalyticsEvent.EXPERIMENT_EXPOSED, {test});
-      }, 1000);
-    }
-  };
-
-  // https://shopify.dev/docs/api/customer-privacy#use-an-event-listener
-  useEffect(() => {
-    const handleConsentChange = (event: any) => {
-      const hasAnalyticsAllowed = event?.detail?.analyticsAllowed || false;
-
-      setHasUserConsent(hasAnalyticsAllowed);
-      console.log('Consent collected:', event.detail); // Log the event details if needed
-
-      if (hasAnalyticsAllowed) {
-        publish?.(AnalyticsEvent.EXPERIMENT_EXPOSED, {test: {adf: 124}});
-      }
-    };
-
-    // Add event listener for 'visitorConsentCollected'
-    document.addEventListener('visitorConsentCollected', handleConsentChange);
-
-    return () => {
-      // Clean up the event listener on unmount
-      document.removeEventListener(
-        'visitorConsentCollected',
-        handleConsentChange,
-      );
-    };
-  }, []);
+  const {handleTestExpose, hasUserConsent} = useTestExpose();
 
   return (
     <>
