+++ app/hooks/index.ts
@@ -19,3 +19,4 @@ export * from './usePreviewMode';
 export * from './usePromobar';
 export * from './useRootLoaderData';
 export * from './useSettings';
+export * from './useTestExpose';
